@@ -1,13 +1,16 @@
 import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
-import { SocketIoModule, SocketIoConfig } from 'ng-socket-io';
+import { NgIoModule, NgIoConfig } from 'ng-io';
 
-const config: SocketIoConfig = { url: 'http://localhost:5000/', options: {} }
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {MatButtonModule, MatInputModule, MatFormFieldModule} from '@angular/material';
 
 import { AppComponent } from './app.component';
 import { ChatComponent } from './components/chat/chat.component';
 import {SocketService} from './services/socket.service';
 
+const config: NgIoConfig = { url: 'http://localhost:5000', options: {} };
 
 @NgModule({
   declarations: [
@@ -16,7 +19,12 @@ import {SocketService} from './services/socket.service';
   ],
   imports: [
     BrowserModule,
-    SocketIoModule.forRoot(config)
+    FormsModule,
+    NgIoModule.forRoot(config),
+    BrowserAnimationsModule,
+    MatButtonModule,
+    MatInputModule,
+    MatFormFieldModule
   ],
   providers: [
     SocketService
